Add AGC config for BirdsEye, add zst support for Ettus and Lime, refactor for ease of maintenance and add tests.
