+++ Dockerfile.api
@@ -1,6 +1,6 @@
 FROM iqtlabs/gamutrf:latest as gamutrf-builder
 RUN apt-get update && apt-get install -y --no-install-recommends software-properties-common && \
-    apt-get -y --no-install-recommends install build-essential cmake g++ gcc git libboost-all-dev libusb-1.0-0-dev libusb-1.0-0 zlib1g-dev make libev-dev zlib1g pkg-config python3-numpy libuhd-dev libjpeg-dev && apt-get clean && rm -rf /var/lib/apt/lists/*
+    apt-get -y --no-install-recommends install build-essential cmake g++ gcc git libboost-all-dev libusb-1.0-0-dev libusb-1.0-0 zlib1g-dev make libev-dev zlib1g pkg-config python3-numpy libuhd-dev libjpeg-dev zstd uhd-host && apt-get clean && rm -rf /var/lib/apt/lists/*
 WORKDIR /root
 RUN git clone https://github.com/anarkiwi/mt_rx_samples_to_file -b v1.0.3
 WORKDIR /root/mt_rx_samples_to_file/build
